requirements.txt
asgiref==3.8.1
Comparable==1.0
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
distlib==0.3.8
Django==5.0.4
et-xmlfile==1.1.0
filelock==3.13.4
fonttools==4.51.0
inflect==7.2.1
keyboard==0.13.5
kiwisolver==1.4.5
matplotlib==3.8.4
more-itertools==10.2.0
MouseInfo==0.1.3
numpy==1.26.4
openpyxl==3.1.5
packaging==24.0
pandas==2.2.2
pdfminer==20191125
pillow==10.3.0
platformdirs==4.2.1
plotly==5.22.0
PyAutoGUI==0.9.54
pycryptodome==3.20.0
pygame==2.5.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.1
scipy==1.13.1
self==2020.12.3
simplex==0.2.3
six==1.16.0
sqlparse==0.5.0
Student==0.0.1
tenacity==8.3.0
typeguard==4.2.1
typing_extensions==4.11.0
tzdata==2024.1
virtualenv==20.26.0
